"Разбиение файла main.js на модули"
