+++ js/check-line-length.js
@@ -1,0 +1,5 @@
+const checkLineLength = (line, maxLength) => {
+  return line.length < maxLength;
+};
+
+export{checkLineLength};
